requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.2.2
numpy==1.26.4
gspread==6.1.2
google-auth==2.30.0
google-auth-oauthlib==1.2.0